Backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
cloudinary==1.37.0
python-dotenv==1.0.0
ultralytics==8.0.196
numpy==1.24.3
Pillow==9.5.0
requests==2.31.0
gunicorn==21.2.0 